pyproject.toml
[project]
name = "codematics"
version = "0.1.0"
description = "Automated CI/CD pipeline for deploying Jupyter Notebooks to Google Cloud Run"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Ifihanagbara Olusheye", email = "victoriaolusheye@gmail.com" }
]
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.121.2",
    "google-auth>=2.43.0",
    "google-cloud-aiplatform>=1.72.0",
    "google-cloud-build>=3.33.0",
    "google-cloud-run>=0.12.0",
    "google-cloud-storage>=3.5.0",
    "google-cloud-monitoring>=2.14.0",
    "google-cloud-logging>=3.0.0",
    "grpc-google-iam-v1>=0.14.3",
    "jupyter>=1.1.1",
    "nbconvert>=7.16.6",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
